Extracts an inner class into a top-level one to support readability.
