+++ src/main/java/sirius/kernel/nls/Babelfish.java
@@ -220,85 +220,4 @@ public class Babelfish {
             throw Exceptions.handle().withSystemErrorMessage("Missing translations found: %s", missing).handle();
         }
     }
-
-    /*
-     * Disables the cache for properties files to enable reloading
-     * and loads UTF-8 stored properties
-     */
-    private static class NonCachingUTF8Control extends ResourceBundle.Control {
-
-        @Override
-        public long getTimeToLive(String baseName, Locale locale) {
-            return ResourceBundle.Control.TTL_DONT_CACHE;
-        }
-
-        @Override
-        public ResourceBundle newBundle(String baseName,
-                                        Locale locale,
-                                        String format,
-                                        ClassLoader loader,
-                                        boolean reload)
-                throws IllegalAccessException, InstantiationException, IOException {
-            // The below is a copy of the default implementation.
-            String bundleName = toBundleName(baseName, locale);
-            String resourceName = toResourceName(bundleName, "properties");
-            ResourceBundle bundle = null;
-            InputStream stream = null;
-            if (reload) {
-                URL url = loader.getResource(resourceName);
-                if (url != null) {
-                    URLConnection connection = url.openConnection();
-                    if (connection != null) {
-                        connection.setUseCaches(false);
-                        stream = connection.getInputStream();
-                    }
-                }
-            } else {
-                stream = loader.getResourceAsStream(resourceName);
-            }
-            if (stream != null) {
-                // Only this line is changed to make it to read properties files as UTF-8.
-                try (InputStreamReader reader = new InputStreamReader(stream, "UTF-8")) {
-                    bundle = new PropertyResourceBundle(reader);
-                } finally {
-                    stream.close();
-                }
-            }
-            return bundle;
-        }
-    }
-
-    private void importProperties(String baseName, String lang) {
-        ResourceBundle bundle = ResourceBundle.getBundle(baseName + "_" + lang, CONTROL);
-        translationsWriteLock.lock();
-        try {
-            Map<String, Translation> copy = new TreeMap<>(translationMap);
-            for (String key : bundle.keySet()) {
-                String value = bundle.getString(key);
-                importProperty(copy, lang, baseName, key, value);
-            }
-            translationMap = copy;
-        } finally {
-            translationsWriteLock.unlock();
-        }
-    }
-
-    private static void importProperty(Map<String, Translation> modifyableTranslationsCopy,
-                                       String lang,
-                                       String file,
-                                       String key,
-                                       String value) {
-        Translation entry = modifyableTranslationsCopy.get(key);
-        if (entry != null && entry.isAutocreated()) {
-            // The entry was previously auto created - therefore we now replace it with our contents. This is a rare
-            // case where a static initialize might fetch a property before the initialization of this class.
-            entry = null;
-        }
-        if (entry == null) {
-            entry = new Translation(key);
-            entry.setFile(file);
-            modifyableTranslationsCopy.put(key, entry);
-        }
-        entry.addTranslation(lang, value, file.startsWith("configurations"));
-    }
 }
